chore(dms): fix pmd
