+++ dms-service/src/test/java/de/muenchen/oss/swim/dms/application/usecase/helper/DmsHelperTest.java
@@ -1,0 +1,5 @@
+package de.muenchen.oss.swim.dms.application.usecase.helper;
+
+class DmsHelperTest {
+
+}
